chore(klaviyo): remove dev mock-fallback layer and add tenant update API
- Drop KlaviyoService's tryReal()/forceMock() machinery, the in-memory
  klaviyoMockData layer, the ?mock=1 query toggle, and the corresponding
  "Showing mock data" banner strings. The OMS endpoints are deployed,
  so failures should surface as real errors instead of silently falling
  back to mock data.
- Inline the previously imported types (CommGatewayAuth, etc.) so
  KlaviyoService is self-contained.
- Add KlaviyoService.updateSystemMessageRemote() backed by the existing
  PUT /oms/systemMessageRemotes/{id} entity-store endpoint, so the UI
  can write back changes to UNIGATE_CONFIG.
- Update documentation/klaviyo-api-contracts.md to drop the Mock mode
  section.
