+++ src/locales/en.json
@@ -772,8 +772,6 @@
   "Reload this page": "Reload this page",
   "Once UNIGATE_CONFIG exists on the OMS, refresh and you'll be able to add a Klaviyo connection here.": "Once UNIGATE_CONFIG exists on the OMS, refresh and you'll be able to add a Klaviyo connection here.",
   "Check again": "Check again",
-  "Showing mock data": "Showing mock data",
-  "The OMS instance has not deployed the Klaviyo endpoints yet. The UI is running against in-memory mock data so you can preview the experience.": "The OMS instance has not deployed the Klaviyo endpoints yet. The UI is running against in-memory mock data so you can preview the experience.",
   "Untitled connection": "Untitled connection",
   "API key": "API key",
   "Email events": "Email events",
